requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click>=8.0,<9
easyocr==1.7.2
# 0.128.x: compatible con Python 3.9 y 3.11. Para solo 3.10+ puedes usar 0.135.x
fastapi==0.128.8
filelock==3.25.0
fsspec==2026.2.0
h11==0.16.0
httptools==0.7.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
pillow-heif>=0.13.0
pyclipper==1.4.0
pydantic==2.12.5
pytesseract==0.3.13
pydantic_core==2.41.5
python-bidi==0.6.7
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.1
setuptools==82.0.0
shapely==2.1.2
starlette>=0.40.0,<1.0.0
sympy==1.14.0
tifffile==2026.3.3
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
openai>=1.0.0
requests==2.32.3
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
